Implements classifiers configuration report fetch method

+++ eca-report/src/main/java/com/ecaservice/report/model/ClassifiersConfigurationBean.java
@@ -45,7 +45,7 @@ public class ClassifiersConfigurationBean {
     /**
      * Classifiers options count associated with configuration
      */
-    private long classifiersOptionsCount;
+    private int classifiersOptionsCount;
 
     /**
      * Classifier options list associated with configuration
